Add blockchain client
The client itself is an interface with eth blockchain methods.
It has an internal request manager that interacts with the blockchain via websocket.
Also, it came along with two default eth subscriptions: newHeads and logs (in order to confirm transactions).

Currently, the supported methods are: eth_blockNumber, eth_chainId, net_listening, eth_getTransactionCount, eth_getTransactionReceipt,
eth_sendRawTransaction and eth_getTransactionByHash.
Besides that, it provides interfaces to eth_call and to others subscriptions.

Co-authored-by: Renato Dinhani <renato.dinhani@cloudwalk.io>